MINIFI-279: Makes test classes naming consistent.
